[EDIFIKANA] #388 updating tests, changing Long to Double for currency values,

+++ edifikana/back-end/src/main/kotlin/com/cramsan/edifikana/server/datastore/PaymentRecordDatastore.kt
@@ -20,7 +20,7 @@ interface PaymentRecordDatastore {
         unitId: UnitId,
         paymentType: PaymentType,
         periodMonth: LocalDate,
-        amountDue: Long?,
+        amountDue: Double?,
         dueDate: LocalDate?,
         recordedBy: UserId?,
         notes: String?,
@@ -46,7 +46,7 @@ interface PaymentRecordDatastore {
      */
     suspend fun updatePaymentRecord(
         paymentRecordId: PaymentRecordId,
-        amountPaid: Long?,
+        amountPaid: Double?,
         paidDate: LocalDate?,
         status: PaymentStatus?,
         notes: String?,
